move NUMERICSTRING certificate test to test_openssl & make it more specific
