+++ src/cryptography/hazmat/backends/openssl/decode_asn1.py
@@ -710,7 +710,7 @@ def _asn1_string_to_utf8(backend, asn1_string):
     buf = backend._ffi.new("unsigned char **")
     res = backend._lib.ASN1_STRING_to_UTF8(buf, asn1_string)
     if res == -1:
-        raise TypeError(
+        raise ValueError(
             "Unsupported ASN1 string type. Type: {0}".format(asn1_string.type)
         )
